+++ beacon/execution/logs/process_logs_fuzz_test.go
@@ -32,8 +32,8 @@ import (
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	loghandler "github.com/itsdevbear/bolaris/beacon/execution/logs"
-	logmocks "github.com/itsdevbear/bolaris/beacon/execution/logs/mocks"
-	"github.com/itsdevbear/bolaris/runtime/logs"
+	"github.com/itsdevbear/bolaris/beacon/staking/logs"
+	logmocks "github.com/itsdevbear/bolaris/beacon/staking/logs/mocks"
 	consensusv1 "github.com/itsdevbear/bolaris/types/consensus/v1"
 	"github.com/stretchr/testify/require"
 )
