+++ beacon/staking/logs/mocks/factory.go
